Organize pages and update their names to what we call them now
