+++ earn/src/components/advanced/BorrowGraphPlaceholder.tsx
