+++ earn/src/components/advanced/GlobalStatsTable.tsx
