+++ earn/src/components/advanced/SmartWalletButton.tsx
