+++ earn/src/components/advanced/modal/NewSmartWalletModal.tsx
