+++ earn/src/components/advanced/modal/tab/AddUniswapNFTAsCollateralTab.tsx
