+++ earn/src/components/common/HealthBar.spec.tsx
