+++ earn/src/components/common/HealthBar.tsx
