+++ earn/src/components/markets/modal/SupplyModal.tsx
@@ -19,7 +19,7 @@ import { Address, useAccount, useBalance, useContractWrite, usePrepareContractWr
 
 import { ChainContext } from '../../../App';
 import { TokenIconsWithTooltip } from '../../common/TokenIconsWithTooltip';
-import { SupplyTableRow } from '../SupplyTable';
+import { SupplyTableRow } from '../supply/SupplyTable';
 
 const SECONDARY_COLOR = 'rgba(130, 160, 182, 1)';
 const TERTIARY_COLOR = '#4b6980';
