+++ earn/src/components/markets/monitor/InfoGraphTooltip.tsx
