+++ earn/src/components/markets/supply/YieldTokenIcons.tsx
